front/requirements.txt
aioice==0.9.0
aiortc==1.5.0
altair==5.0.1
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bokeh==3.1.1
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
compress-pickle==2.1.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dnspython==2.4.2
Faker==19.3.1
favicon==0.7.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.32
google-crc32c==1.5.0
htbuilder==0.6.1
hydralit==1.0.14
hydralit-components==1.0.10
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
numpy==1.24.4
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
protobuf==4.24.1
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.0
pyee==11.0.0
Pygments==2.16.1
pylibsrtp==0.8.0
pymdown-extensions==10.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.6.1
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
st-annotated-text==4.0.0
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
streamlit-webrtc==0.47.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
websocket-client==1.6.2
wheel==0.38.4
xyzservices==2023.7.0
zipp==3.16.2
